--- publicaciones.py
@@ -1,35 +1,0 @@
-#!/usr/bin/python
-# -*- encoding: utf-8 -*-
-
-#######################################
-# ###     Raúl Caro Pastorino     ### #
-## ##                             ## ##
-### # https://github.com/fryntiz/ # ###
-## ##                             ## ##
-# ###       www.fryntiz.es        ### #
-#######################################
-
-##############################
-##    Importar Librerías    ##
-##############################
-
-
-class publicaciones:
-    LINEA_ACTUAL = 0
-    documento_abierto = ''
-    TOTAL_LINEAS = 0
-
-    def __init__(self, archivo_publicaciones):
-        print('Creando Objeto de publicación')
-
-    def comprobar_linea(linea):
-        print('Comprobando línea recibida')
-
-    def publicacion_actual(linea):
-        print('La línea actual es')
-
-        # return string con la línea
-
-#Comprueba la validez de la próxima línea y se posiciona sobre ella
-    def siguiente_linea():
-        print('Estableciéndo contador en la próxima línea')
